Get rid of trait
